pyproject.toml
[project]
name = "redis-social-sentiment-trader"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "redis>=5.0.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "matplotlib>=3.8.0",
  "yfinance>=0.2.30",
  "tqdm>=4.66.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.0.0",
  "tenacity>=8.2.0"
]

[tool.setuptools]
package-dir = {"" = "src"}
